Move utility functions to Utilities.ts
