+++ resources/assets/ts/ToggleAction.ts
@@ -1,5 +1,5 @@
 import * as $ from "jquery"
-import {safeSetLocalStorage} from "./EventView";
+import {safeSetLocalStorage} from "./Utilities";
 
 export interface ToggleAction {
     (isActive: boolean): void;
@@ -9,5 +9,5 @@ export const makeLocalStorageAction = (key: string, active, inactive) => (
     (isActive: boolean) => safeSetLocalStorage(key, isActive ? inactive : active)
 );
 export const makeClassToggleAction = (selector: string | JQuery, style: string, isActiveOnToggle) => (
-    (isActive: boolean) => $(selector).toggleClass(style, isActive ? ! isActiveOnToggle : isActiveOnToggle)
+    (isActive: boolean) => $(selector).toggleClass(style, isActive ? !isActiveOnToggle : isActiveOnToggle)
 )
